Load and display last event sessions
