+++ app/src/main/java/fr/groggy/racecontrol/tv/ui/home/HomeFragment.kt
@@ -1,6 +1,7 @@
 package fr.groggy.racecontrol.tv.ui.home
 
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -11,11 +12,16 @@ import androidx.core.view.updateLayoutParams
 import androidx.fragment.app.viewModels
 import androidx.leanback.app.RowsSupportFragment
 import androidx.leanback.widget.*
+import androidx.lifecycle.asLiveData
+import androidx.lifecycle.lifecycleScope
 import dagger.hilt.android.AndroidEntryPoint
 import fr.groggy.racecontrol.tv.R
 import fr.groggy.racecontrol.tv.f1tv.Archive
 import fr.groggy.racecontrol.tv.ui.season.archive.SeasonArchiveActivity
+import fr.groggy.racecontrol.tv.ui.season.browse.Season
 import fr.groggy.racecontrol.tv.ui.season.browse.SeasonBrowseActivity
+import fr.groggy.racecontrol.tv.ui.season.browse.Session
+import fr.groggy.racecontrol.tv.ui.session.SessionCardPresenter
 import org.threeten.bp.Year
 
 @Keep
@@ -28,6 +34,7 @@ class HomeFragment : RowsSupportFragment(), OnItemViewClickedListener {
     }
     private val archivesAdapter = ArrayObjectAdapter(listRowPresenter)
     private var imageView: ImageView? = null
+    private val currentYear = Year.now().value
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -43,7 +50,8 @@ class HomeFragment : RowsSupportFragment(), OnItemViewClickedListener {
     ): View? {
         val view = super.onCreateView(inflater, container, savedInstanceState)
         view?.updateLayoutParams<ViewGroup.MarginLayoutParams> {
-            val dimensionPixelSize = inflater.context.resources.getDimensionPixelSize(R.dimen.lb_browse_rows_fading_edge)
+            val dimensionPixelSize =
+                inflater.context.resources.getDimensionPixelSize(R.dimen.lb_browse_rows_fading_edge)
             val horizontalMargin = -dimensionPixelSize * 2 - 4
 
             leftMargin = horizontalMargin
@@ -58,8 +66,6 @@ class HomeFragment : RowsSupportFragment(), OnItemViewClickedListener {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        val currentYear = Year.now().value
-
         imageView?.requestFocus()
         imageView?.setOnClickListener {
             val activity = SeasonBrowseActivity.intent(requireContext(), Archive(currentYear))
@@ -73,8 +79,26 @@ class HomeFragment : RowsSupportFragment(), OnItemViewClickedListener {
     private fun buildRowsAdapter() {
         val viewModel: HomeViewModel by viewModels()
 
-        archivesAdapter.add(getArchiveRow(viewModel))
-//        archivesAdapter.add(ListRow(HeaderItem("Documentations"), listRowAdapter))
+        lifecycleScope.launchWhenStarted {
+            viewModel.getCurrentSeason(Archive(currentYear)).asLiveData()
+                .observe(viewLifecycleOwner, ::onUpdatedSeason)
+        }
+    }
+
+    private fun onUpdatedSeason(season: Season) {
+        val viewModel: HomeViewModel by viewModels()
+        val event = season.events.filter { it.sessions.isNotEmpty() }[0]
+        val listRowAdapter = ArrayObjectAdapter(SessionCardPresenter())
+        listRowAdapter.setItems(event.sessions, Session.diffCallback)
+
+        if (archivesAdapter.size() == 0) {
+            archivesAdapter.add(ListRow(HeaderItem(event.name + " " + currentYear), listRowAdapter))
+            archivesAdapter.add(getArchiveRow(viewModel))
+        } else {
+            val listRow: ListRow = archivesAdapter.get(0) as ListRow
+            (listRow.adapter as ArrayObjectAdapter).setItems(event.sessions, Session.diffCallback)
+        }
+
     }
 
     private fun getArchiveRow(viewModel: HomeViewModel): ListRow {
@@ -83,7 +107,12 @@ class HomeFragment : RowsSupportFragment(), OnItemViewClickedListener {
 
         val listRowAdapter = ArrayObjectAdapter(HomeItemPresenter())
         listRowAdapter.setItems(archives, null)
-        listRowAdapter.add(HomeItem(HomeItemType.ARCHIVE_ALL, resources.getString(R.string.home_all)))
+        listRowAdapter.add(
+            HomeItem(
+                HomeItemType.ARCHIVE_ALL,
+                resources.getString(R.string.home_all)
+            )
+        )
 
         return ListRow(HeaderItem(resources.getString(R.string.home_archive)), listRowAdapter)
     }
